chore: refactor dir architecture
