+++ apps/web/src/query/core/query-key/use-service-query-key.md
