+++ apps/web/src/query/resource-query/reference-model/composables/_internal/use-batched-reference-fetch.ts
@@ -3,7 +3,7 @@ import {
 } from 'vue';
 
 import { referenceQueryClient as queryClient } from '@/query/clients';
-import type { QueryKeyArray } from '@/query/query-key/types/query-key-type';
+import type { QueryKeyArray } from '@/query/core/query-key/types/query-key-type';
 import type { ReferenceConfig } from '@/query/resource-query/reference-model/types/reference-type';
 import { createResourceIdResolver } from '@/query/resource-query/reference-model/utils/reference-helper';
 import { useResourceInfo } from '@/query/resource-query/shared/composable/use-resource-info';
