+++ apps/web/src/query/service-query/use-scoped-infinite-query.ts
@@ -1,5 +1,5 @@
 /**
- * useScopedQuery - A custom wrapper around `useQuery` to enforce scope-based API fetching.
+ * useScopedInfiniteQuery - A custom wrapper around `useInfiniteQuery` to enforce scope-based API fetching.
  *
  * ## Why this hook exists?
  * This hook was created to integrate **scope-based API access control** with Vue Query.
@@ -7,23 +7,22 @@
  * Additionally, it automatically handles loading states and prevents unnecessary queries.
  *
  * ## Functionality
- * - Extends `useQuery` with **grant scope validation**.
+ * - Extends `useInfiniteQuery` with **grant scope validation**.
  * - Runs queries only when **the user's scope is valid** and the app is **ready**.
  * - Uses Vue's **reactivity** to dynamically compute the `enabled` state.
  * - Supports both **static and reactive `enabled` values**.
  *
  * ## Parameters:
- * - `options`: Standard **Vue Query options** (`UseQueryOptions`).
+ * - `options`: Standard **Vue Query options** (`UseInfiniteQueryOptions`).
  * - `requiredScopes`: A list of **required grant scopes** to determine if the query should execute.
  *
- * ## Supported Query Options:
- * - `queryKey`, `queryFn`, `select`, `initialData`, `staleTime`, `enabled` (static or reactive)
- *
  * ## Example:
- * const query = useScopedQuery(
+ * const query = useScopedInfiniteQuery(
  *   {
  *     queryKey: ['dashboard', dashboardId],
  *     queryFn: () => fetchDashboardData(dashboardId),
+ *     initialPageParam: 1,
+ *     getNextPageParam: (lastPage, allPages) => {}
  *     enabled: computed(() => isUserAuthorized.value),
  *   },
  *   ['DOMAIN', 'WORKSPACE']
@@ -34,12 +33,16 @@ import type { MaybeRef } from '@vueuse/core';
 import { toValue } from '@vueuse/core';
 import { computed, type ComputedRef } from 'vue';
 
+import type {
+    InfiniteData,
+    QueryKey,
+} from '@tanstack/vue-query';
 import {
-    useQuery, type UseQueryOptions, type UseQueryReturnType,
+    useInfiniteQuery, type UseInfiniteQueryOptions, type UseInfiniteQueryReturnType,
 } from '@tanstack/vue-query';
 
 import type { GrantScope } from '@/api-clients/identity/token/schema/type';
-import type { QueryKeyArray } from '@/query/query-key/types/query-key-type';
+import type { QueryKeyArray } from '@/query/core/query-key/types/query-key-type';
 
 import { useAppContextStore } from '@/store/app-context/app-context-store';
 import { useAuthorizationStore } from '@/store/authorization/authorization-store';
@@ -48,15 +51,15 @@ import { useAuthorizationStore } from '@/store/authorization/authorization-store
 type ScopedEnabled = MaybeRef<boolean>;
 
 
-export const useScopedQuery = <TQueryFnData, TError = unknown, TData = TQueryFnData>(
-    options: UseQueryOptions<TQueryFnData, TError, TData>,
+export const useScopedInfiniteQuery = <TQueryFnData, TError = unknown, TData = InfiniteData<TQueryFnData>, TQueryKey extends QueryKey = QueryKey, TPageParam = unknown>(
+    options: UseInfiniteQueryOptions<TQueryFnData, TError, TData, TQueryFnData, TQueryKey, TPageParam>,
     requiredScopes: [GrantScope, ...GrantScope[]],
-): UseQueryReturnType<TData, TError> => {
+): UseInfiniteQueryReturnType<TData, TError> => {
     // [Dev Warning] This query is missing `requiredScopes`.
     // All scoped queries must explicitly define at least one valid scope for clarity and safety.
     if (import.meta.env.DEV && (!requiredScopes || requiredScopes.length === 0)) {
         _warnOncePerTick(() => {
-            console.warn('[useScopedQuery] `requiredScopes` is missing or empty.', {
+            console.warn('[useScopedInfiniteQuery] `requiredScopes` is missing or empty.', {
                 queryKey: _extractQueryKey((options as any).queryKey),
                 suggestion: 'Pass at least one valid scope like [\'DOMAIN\'], [\'WORKSPACE\'], etc.',
             });
@@ -87,7 +90,7 @@ export const useScopedQuery = <TQueryFnData, TError = unknown, TData = TQueryFnD
         const currentScope = currentGrantScope.value;
         if (isAppReady.value && currentScope && toValue(rawEnabled) && !requiredScopes.includes(currentScope)) {
             _warnOncePerTick(() => {
-                console.warn('[useScopedQuery] Invalid requiredScopes for current scope:', {
+                console.warn('[useScopedInfiniteQuery] Invalid requiredScopes for current scope:', {
                     queryKey: _extractQueryKey((options as any).queryKey),
                     requiredScopes,
                     currentScope,
@@ -97,7 +100,7 @@ export const useScopedQuery = <TQueryFnData, TError = unknown, TData = TQueryFnD
         }
     }
 
-    return useQuery<TQueryFnData, TError, TData>({
+    return useInfiniteQuery<TQueryFnData, TError, TData, TQueryKey, TPageParam>({
         ...options,
         enabled: queryEnabled,
     });
@@ -116,7 +119,7 @@ const _getCallerKey = (): string => {
 
         const caller = stack.find((line, i) => i > 1
             && (line.includes('.ts') || line.includes('.vue'))
-            && !line.includes('use-scoped-query'));
+            && !line.includes('use-scoped-infinite-query'));
 
         return caller?.trim() ?? 'UNKNOWN_CALLSITE';
     } catch {
